sched/fair: Don't allow migrating boosted tasks to small cores
We want boosted tasks to run on big cores. But CAF's load balancer
changes do not account for SchedTune boosting, so this allows for
boosted tasks to be migrated to a suboptimal core. Let's mitigate
this by setting the LBF_IGNORE_BIG_TASKS for tasks migrating from
a larger capacity core to a smaller one and to check if the task
is SchedTune boosted. If both are true, do not migrate this task.

Change-Id: Idf51ca3cc2e62dd8ff18ab5f46a6c5bfa7696eca
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>